add Login to AdminService
