+++ bolt/main.go
@@ -1,10 +1,15 @@
 package bolt
 
 import (
+	"errors"
 	"log"
+	"regexp"
+
+	"golang.org/x/crypto/bcrypt"
 
 	"github.com/cayleygraph/cayley"
 	"github.com/cayleygraph/cayley/graph"
+	"github.com/cayleygraph/cayley/schema"
 )
 
 // init and open
@@ -24,4 +29,26 @@ func Open(dbFile string) (*cayley.Handle, error) {
 // AdminService represents a PostgreSQL implementation of myapp.UserService.
 type AdminService struct {
 	Store *cayley.Handle
+}
+
+func validateEmail(email string) error {
+	var ErrBadFormat = errors.New("invalid email format")
+	var emailRegexp = regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
+
+	if !emailRegexp.MatchString(email) {
+		return ErrBadFormat
+	}
+	return nil
+}
+
+func hashPassword(password string) (string, error) {
+	bytes, err := bcrypt.GenerateFromPassword([]byte(password), 14)
+	return string(bytes), err
+}
+
+func insert(store *cayley.Handle, o interface{}) error {
+	qw := graph.NewWriter(store)
+	defer qw.Close() // don't forget to close a writer; it has some internal buffering
+	_, err := schema.WriteAsQuads(qw, o)
+	return err
 }
